refactor: VC 역활에 대한 코드 분기처리
