+++ CloudNotes/CloudNotes/Controller/SplitViewController.swift
@@ -10,13 +10,15 @@ class SplitViewController: UISplitViewController, TextSeparatable {
     // MARK: Property
     private let memoListViewController = MemoListTableViewController()
     private let memoDetailViewController = MemoDetailViewController()
-    
+    private var dataSource: MemoListDiffableDataSource?
     // MARK: View LifeCycle
     override func viewDidLoad() {
         super.viewDidLoad()
+        delegate = self
         setupChildViewController()
         setupSplitViewDisPlayMode()
-        delegate = self
+        makeTableViewDiffableDataSource()
+        CoreDataCloudMemo.shared.perforFetchCloudMemo()
     }
     
 }
@@ -54,6 +56,11 @@ extension SplitViewController: MemoListDelegate {
         composeViewController.delegate = self
         present(UINavigationController(rootViewController: composeViewController), animated: true, completion: nil)
     }
+    
+    func didTapDeleteButton(at indexPath: IndexPath) {
+        let currentObject = CoreDataCloudMemo.shared.fetchedController.object(at: indexPath)
+        CoreDataCloudMemo.shared.deleteItem(object: currentObject)
+    }
 }
 
 extension SplitViewController: ComposeTextViewControllerDelegate {
@@ -63,4 +70,20 @@ extension SplitViewController: ComposeTextViewControllerDelegate {
                                                body: texts.body,
                                                lastModifier: Date())
     }
+}
+
+// MARK: - Diffable DataSource
+extension SplitViewController {
+    private func makeTableViewDiffableDataSource() {
+        dataSource = MemoListDiffableDataSource(tableView: memoListViewController.tableView) { tableView, indexPath, memo in
+            
+            guard let cell = tableView.dequeueReusableCell(withIdentifier: MemoListTableViewCell.identifier, for: indexPath) as? MemoListTableViewCell else { fatalError() }
+            
+            cell.accessoryType = .disclosureIndicator
+            cell.configure(with: memo)
+            
+            return cell
+        }
+    }
+    
 }
